external datasets, uploads and provenance

+++ services/src/datasets/upload.rs
@@ -8,6 +8,8 @@ use crate::{
 };
 use async_trait::async_trait;
 use geoengine_datatypes::identifier;
+#[cfg(feature = "postgres")]
+use postgres_types::{FromSql, ToSql};
 use serde::{Deserialize, Serialize};
 
 identifier!(UploadId);
@@ -40,10 +42,11 @@ impl Upload {
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
+#[cfg_attr(feature = "postgres", derive(ToSql, FromSql))]
 pub struct FileUpload {
     pub id: FileId,
     pub name: String,
-    pub byte_size: usize,
+    pub byte_size: i64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
